fix type indexing

+++ src/instructions.rs
@@ -39,10 +39,10 @@ struct ADD;
 
 impl Instruction for ADD {
     fn execute(emulator: &mut Emulator, args: InstructionArgs) {
-        let a = emulator.registers[args.get_nibble(1)];
-        let b = emulator.registers[args.get_nibble(2)];
+        let a = emulator.registers[args.get_nibble(1) as usize];
+        let b = emulator.registers[args.get_nibble(2) as usize];
         let c = a + b;
-        emulator.registers[args.get_nibble(3)] = c;
+        emulator.registers[args.get_nibble(3) as usize] = c;
     }
     fn bytes_len() -> i32 {
         2
